requirements.txt
# Core inference engine
# NOTE: install vLLM first with its version-specific wheel index, e.g.:
#   pip install vllm==0.11.2 --extra-index-url https://wheels.vllm.ai/0.11.2/
vllm==0.11.2

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12

# Data & validation
pydantic==2.9.2
pydantic-settings==2.5.2

# Image handling
Pillow==10.4.0
numpy==1.26.4

# Observability
prometheus-client>=0.20
structlog==24.4.0

# Async I/O
aiofiles==24.1.0
httpx==0.27.2
